pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jtap"
version = "0.1.0"
description = "A Python library for probabilistic psychophysical modeling and inference of 2D Mental Physics"
authors = [{ name = "Arijit Dasgupta", email = "arijitdg@mit.edu" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
keywords = ["physics", "probabilistic", "intuitive", "cognitive"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "attrs>=25.4.0",
    "chex==0.1.86",
    "dtaidistance==2.3.13",
    "genjax==0.10.3",
    "imageio[ffmpeg]>=2.37.2",
    "ipyfilechooser==0.6.0",
    "ipykernel>=7.0.1",
    "ipython==8.12.3",
    "ipywidgets==8.1.2",
    "jax==0.5.3",
    "matplotlib==3.8.4",
    "numpy==1.26.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pingouin==0.5.5",
    "plotly==6.2.0",
    "pymunk>=7.2.0",
    "pyyaml>=6.0.3",
    "rerun-sdk>=0.25.0",
    "rich>=14.2.0",
    "scikit-learn==1.7.0",
    "seaborn==0.13.2",
    "statsmodels==0.14.4",
    "tensorflow-probability==0.23.0",
    "termcolor==3.1.0",
    "tqdm>=4.67.1",
]

# Platform-specific dependencies
[project.optional-dependencies]
# For NVIDIA GPU systems with CUDA 12 (default for GPU users)
cuda = [
    "jax[cuda12]==0.5.3",
]
# For CPU-only systems (fallback for systems without NVIDIA GPU)
cpu = [
    "jax[cpu]==0.5.3",
]

[tool.uv]
# Override Rerun's numpy>2.0 requirement due to the conflicts with GenJAX
override-dependencies = ["numpy==1.26.4"]
